move test files next to source
