+++ mock/entry.mock.ts
