+++ tsconfig.json
@@ -70,5 +70,5 @@
     "skipLibCheck": true,                           /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */
   },
-  "exclude": ["./test/**/*"]
+  "exclude": ["./src/**/*.test.ts", "./mock/**/*"]
 }
